Do not pass byte array to transit reader.
The newest version of clj-http assumes that it is always dealing with streams,
but that assumption does not seem to be a problem except when trying to parse
transit using `:as :transit+json`.  In that case clj-http takes the body of the
request and passes it through to `transit/reader`, and when the body is a byte
array (as it is when clj-http-fake is intercepting), then the transit reader
throws an exception.

The fix is to convert the byte array into an input stream during interception.